requirements.txt
streamlit
huggingface_hub[hf_xet]
torch
transformers
pypdf
python-docx
python-pptx
